Move SyntheticNodes to top-level construction package and rename
